# docs/swagger.yaml
basePath: /api
definitions:
...
  common.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  employee.CreateEmployeeRequest:
    properties:
      department_id:
        type: string
      designation:
        type: string
      joining_date:
        type: string
      salary:
        type: number
      status:
        type: string
      user_id:
        type: string
    required:
    - department_id
    - designation
    - joining_date
    - user_id
    type: object
  employee.UpdateEmployeeRequest:
    properties:
      department_id:
        type: string
      designation:
        type: string
      joining_date:
        description: Added (or use time.Time)
        type: string
      salary:
        type: number
      status:
        type: string
      user_id:
        description: Added
        type: string
    type: object
  leave.CreateLeaveRequest:
    properties:
      employee_id:
        type: string
      from_date:
        type: string
      leave_type:
        type: string
      reason:
        type: string
      to_date:
        type: string
    required:
    - employee_id
    - from_date
    - leave_type
    - reason
    - to_date
    type: object
  leave.UpdateLeaveStatusRequest:
    properties:
      remarks:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ravi@example.com
    name: Raviraj Sahu
  description: HRMS Backend API using Go, Gin, GORM & PostgreSQL.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: HRMS API
  version: "1.0"
paths:
...
  /employees:
    get:
      description: Get All employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Employees
      tags:
      - Employee
    post:
      consumes:
      - application/json
      description: Create a new employee
      parameters:
      - description: Employee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Employee
      tags:
      - Employee
  /employees/{id}:
    delete:
      description: Delete Employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Employee
      tags:
      - Employee
    get:
      description: Get All employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Employee By ID
      tags:
      - Employee
    put:
      consumes:
      - application/json
      description: Update Employee Details
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Employee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Employee
      tags:
      - Employee
  /leaves:
    get:
      description: Get All Leaves
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Leaves
      tags:
      - Leave
    post:
      consumes:
      - application/json
      description: Create a new leave
      parameters:
      - description: Leave
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/leave.CreateLeaveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Leave
      tags:
      - Leave
  /leaves/{id}:
    delete:
      description: Delete (cancel) a leave request
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Leave
      tags:
      - Leave
    get:
      description: Get All leave by ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Leave By ID
      tags:
      - Leave
  /leaves/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve an employee leave request
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      - description: Approval Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/leave.UpdateLeaveStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve Leave
      tags:
      - Leave
  /leaves/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject an employee leave request
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      - description: Reject Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/leave.UpdateLeaveStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Reject Leave
      tags:
      - Leave
  /leaves/employee/{employee_id}:
    get:
      description: Get all leave requests for a specific employee
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Leaves By Employee
      tags:
      - Leave
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"